--- legend-engine-executionPlan-execution-store-relational-connection-tests-api/src/main/resources/META-INF/services/org.finos.legend.engine.plan.execution.stores.relational.connection.tests.api.ConnectionTestManager
@@ -1,2 +1,0 @@
-org.finos.legend.engine.plan.execution.stores.relational.connection.tests.api.connectionTestManagers.H2ConnectionTestManager
-org.finos.legend.engine.plan.execution.stores.relational.connection.tests.api.connectionTestManagers.SqlServerConnectionTestManager
