+++ legend-engine-executionPlan-execution-store-relational-connection-tests-api/src/main/resources/META-INF/services/org.finos.legend.engine.plan.execution.stores.relational.connection.tests.api.DynamicTestConnection
@@ -1,0 +1,2 @@
+org.finos.legend.engine.plan.execution.stores.relational.connection.tests.api.dynamicTestConnections.H2AlloyServer
+org.finos.legend.engine.plan.execution.stores.relational.connection.tests.api.dynamicTestConnections.SqlServerTestContainers
