+++ legend-engine-server/src/test/resources/org/finos/legend/engine/server/test/userTestConfig_withSqlServerTestConnection.json
@@ -76,7 +76,7 @@
       "_type": "legendDefault"
     }
   },
-  "testConnectionsToEnable" : ["h2", "sqlServer"],
+  "testConnectionsToEnable" : ["sqlServer"],
   "dynamicTestConnectionCreators": {
     "h2": "org.finos.legend.engine.plan.execution.stores.relational.connection.tests.api.dynamicTestConnections.H2AlloyServer",
     "sqlServer" : "org.finos.legend.engine.plan.execution.stores.relational.connection.tests.api.dynamicTestConnections.SqlServerTestContainers"
